+++ core/serverless/lambda/src/main/kotlin/org/http4k/serverless/ApiGatewayV1.kt
@@ -52,6 +52,7 @@ object ApiGatewayV1AwsHttpAdapter : AwsHttpAdapter<Map<String, Any>, Map<String,
     override fun invoke(resp: Response) = mapOf(
         "statusCode" to resp.status.code,
         "headers" to resp.headers.toMap(),
+        "multiValueHeaders" to resp.headers.groupBy({ it.first }, { it.second }),
         "body" to resp.body.payload.base64Encode(),
         "isBase64Encoded" to true
     )