+++ core/serverless/lambda/src/test/kotlin/org/http4k/serverless/ApiGatewayRestLambdaFunctionTest.kt
@@ -38,6 +38,7 @@ class ApiGatewayRestLambdaFunctionTest {
                     mapOf(
                         "statusCode" to 200,
                         "headers" to mapOf("a" to "b"),
+                        "multiValueHeaders" to mapOf("a" to listOf("b")),
                         "body" to "hello there",
                         "isBase64Encoded" to false
                     )